Correctly format `unsafe` expressions.
