+++ Tests/SwiftFormatTests/PrettyPrint/DoStmtTests.swift
@@ -43,6 +43,19 @@ final class DoStmtTests: PrettyPrintTestCase {
     assertPrettyPrintEqual(input: input, expected: expected, linelength: 25)
   }
 
+  func testNestedDo() {
+    // Avoid regressions in the case where a nested `do` block was getting shifted all the way left.
+    let input = """
+      func foo() {
+        do {
+          bar()
+          baz()
+        }
+      }
+      """
+    assertPrettyPrintEqual(input: input, expected: input + "\n", linelength: 45)
+  }
+
   func testLabeledDoStmt() {
     let input = """
       someLabel:do {
@@ -101,5 +114,204 @@ final class DoStmtTests: PrettyPrintTestCase {
         """,
       linelength: 25
     )
+  }
+
+  func testDoCatch_noBreakBeforeCatch() {
+    let input =
+      """
+      do { try thisFuncMightFail() } catch error1 { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") } catch error2 { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") } catch error2(let someVar) { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") }
+      catch error2(let someVar) {
+        print(someVar)
+        print("Don't do it!")
+      }
+      do { try thisFuncMightFail() } catch is ABadError{ print("Nope") }
+      """
+
+    let expected =
+      """
+      do {
+        try thisFuncMightFail()
+      } catch error1 { print("Nope") }
+      do {
+        try thisFuncMightFail()
+      } catch error1 {
+        print("Nope")
+      } catch error2 { print("Nope") }
+      do {
+        try thisFuncMightFail()
+      } catch error1 {
+        print("Nope")
+      } catch error2(let someVar) {
+        print("Nope")
+      }
+      do {
+        try thisFuncMightFail()
+      } catch error1 {
+        print("Nope")
+      } catch error2(let someVar) {
+        print(someVar)
+        print("Don't do it!")
+      }
+      do {
+        try thisFuncMightFail()
+      } catch is ABadError { print("Nope") }
+
+      """
+
+    assertPrettyPrintEqual(input: input, expected: expected, linelength: 40)
+  }
+
+  func testDoCatch_breakBeforeCatch() {
+    let input =
+      """
+      do { try thisFuncMightFail() } catch error1 { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") } catch error2 { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") } catch error2(let someVar) { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 { print("Nope") }
+      catch error2(let someVar) {
+        print(someVar)
+        print("Don't do it!")
+      }
+      do { try thisFuncMightFail() } catch is ABadError{ print("Nope") }
+      """
+
+    let expected =
+      """
+      do { try thisFuncMightFail() }
+      catch error1 { print("Nope") }
+      do { try thisFuncMightFail() }
+      catch error1 { print("Nope") }
+      catch error2 { print("Nope") }
+      do { try thisFuncMightFail() }
+      catch error1 { print("Nope") }
+      catch error2(let someVar) {
+        print("Nope")
+      }
+      do { try thisFuncMightFail() }
+      catch error1 { print("Nope") }
+      catch error2(let someVar) {
+        print(someVar)
+        print("Don't do it!")
+      }
+      do { try thisFuncMightFail() }
+      catch is ABadError { print("Nope") }
+
+      """
+
+    var config = Configuration.forTesting
+    config.lineBreakBeforeControlFlowKeywords = true
+    assertPrettyPrintEqual(input: input, expected: expected, linelength: 40, configuration: config)
+  }
+
+  func testCatchWhere_noBreakBeforeCatch() {
+    let input =
+      """
+      do { try thisFuncMightFail() } catch error1 where error1 is ErrorType { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 where error1 is LongerErrorType { print("Nope") }
+      """
+
+    let expected =
+      """
+      do {
+        try thisFuncMightFail()
+      } catch error1 where error1 is ErrorType {
+        print("Nope")
+      }
+      do {
+        try thisFuncMightFail()
+      } catch error1
+        where error1 is LongerErrorType
+      { print("Nope") }
+
+      """
+
+    assertPrettyPrintEqual(input: input, expected: expected, linelength: 42)
+  }
+
+  func testCatchWhere_breakBeforeCatch() {
+    let input =
+      """
+      do { try thisFuncMightFail() } catch error1 where error1 is ErrorType { print("Nope") }
+      do { try thisFuncMightFail() } catch error1 where error1 is LongerErrorType { print("Nope") }
+      """
+
+    let expected =
+      """
+      do { try thisFuncMightFail() }
+      catch error1 where error1 is ErrorType {
+        print("Nope")
+      }
+      do { try thisFuncMightFail() }
+      catch error1
+      where error1 is LongerErrorType {
+        print("Nope")
+      }
+
+      """
+
+    var config = Configuration.forTesting
+    config.lineBreakBeforeControlFlowKeywords = true
+    assertPrettyPrintEqual(input: input, expected: expected, linelength: 42, configuration: config)
+  }
+
+  func testMultipleCatchItems() {
+    let input =
+      """
+      do { try thisMightFail() } catch error1, error2 { print("Nope") }
+      do { try thisMightFail() } catch longErrorType, error2 { print("Nope") }
+      do { try thisMightFail() } catch longErrorTypeName, longErrorType2(let someLongVariable) { print("Nope") }
+      do { try thisMightFail() } catch longErrorTypeName, longErrorType2 as SomeLongErrorType { print("Nope") }
+      do { try thisMightFail() } catch longErrorName where someCondition, longErrorType2 { print("Nope") }
+      do { try thisMightFail() } catch longErrorTypeName, longErrorType2 as SomeLongErrorType where someCondition, longErrorType3 { print("Nope") }
+      """
+
+    let expected =
+      """
+      do {
+        try thisMightFail()
+      } catch error1, error2 {
+        print("Nope")
+      }
+      do {
+        try thisMightFail()
+      } catch longErrorType,
+        error2
+      { print("Nope") }
+      do {
+        try thisMightFail()
+      } catch
+        longErrorTypeName,
+        longErrorType2(
+          let someLongVariable)
+      { print("Nope") }
+      do {
+        try thisMightFail()
+      } catch
+        longErrorTypeName,
+        longErrorType2
+          as SomeLongErrorType
+      { print("Nope") }
+      do {
+        try thisMightFail()
+      } catch longErrorName
+        where someCondition,
+        longErrorType2
+      { print("Nope") }
+      do {
+        try thisMightFail()
+      } catch
+        longErrorTypeName,
+        longErrorType2
+          as SomeLongErrorType
+          where someCondition,
+        longErrorType3
+      { print("Nope") }
+
+      """
+
+    assertPrettyPrintEqual(input: input, expected: expected, linelength: 25)
   }
 }
